Slight color changes + all cells position from int to double

+++ src/elements/Cell.java
@@ -10,7 +10,7 @@ public class Cell {
 	private Cell outCell;
 	
 	// Display
-	private int x,y; // center of the Cell
+	private double x,y; // center of the Cell
 	
 	public Cell() {
 		isOccupied = false;
@@ -40,16 +40,16 @@ public class Cell {
 			System.out.print("[ ]");
 		}
 	}
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	public Cell getPreviousCell() {
